pyproject.toml
[tool.poetry]
name = "pudl-output-differ"
version = "0.1.0"
description = ""
authors = ["Jan Rous <rousik@gmail.com>"]
readme = "README.md"
packages = [{include = "pudl_output_differ", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.3.0"
fsspec = "^2023.9.0"
sqlalchemy = "^2.0.20"
pandas = "^2.1.0"
pydantic-settings = "^2.0.3"
gcsfs = "^2023.9.0"
pyarrow = "^13.0.0"
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"
apsw = "^3.43.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
